Restructured the application
