+++ bin/silverstripe
@@ -1,16 +1,13 @@
 #!/usr/bin/env php
 <?php
 
-// Only use in production
 if (file_exists(__DIR__.'/../../../autoload.php')) {
     require __DIR__.'/../../../autoload.php';
 } else {
     require __DIR__.'/../vendor/autoload.php';
 }
 
-use Symfony\Component\Console\Application;
+use KevinJansen\SilverstripeArtisan\Console\Bootstrap;
 
-$app = new Application('Silverstripe CLI', '0.1');
-$app->add(new \KevinJansen\SilverstripeArtisan\Console\NewProjectCommand);
-
-$app->run();
+$bootstrap = new Bootstrap;
+$bootstrap->getApplication()->run();
